reset game selection on new submit
previously, if you selected a game from a users page to filter by, then switched to another users page who didn't have any mods for that game, no mods would appear. This is due to the selected game variable still being set to the old value. Now, that value is reset upon resubmit.

This also allows other actions to be done on re-submit of the form